requirements.txt
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
plotly==5.24.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
six==1.17.0
tenacity==9.0.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
